requirements.txt
streamlit
openai
python-dotenv
pydantic
pylint
radon
